@@ -2,7 +2,7 @@ package com.readystatesoftware.chuck.internal.support;
 
 import android.app.IntentService;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.readystatesoftware.chuck.internal.data.ChuckContentProvider;
 
