@@ -15,7 +15,7 @@
  */
 package com.readystatesoftware.chuck.internal.support;
 
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 
 public abstract class SimpleOnPageChangedListener implements ViewPager.OnPageChangeListener {
     @Override
