@@ -16,8 +16,8 @@
 package com.readystatesoftware.chuck.internal.ui;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.readystatesoftware.chuck.internal.support.NotificationHelper;
 
